Add files via upload
There are two serious problems that I can pinpoint on why this isn't working, on top of the many other things that must be added.
First the screen gets cleared despite the function to redraw everything, I have no idea why. The recursive function I know for a fact has all the parts being got to, but the problem is... I don't know it's just not drawing.
The other problem is I think sometimes there are more than 1 ellipse being added when you finish it.